Time: 3 ms (81.55%) | Memory: 69.3 MB (90.89%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -1,17 +1,17 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        double sum = 0;
+        int sum = 0;
 
         for (int i = 0; i < k; i++) {
-           sum += nums[i];
+            sum += nums[i];
         }
 
-        double maxSum = sum;
+        int maxSum = sum;
 
         for (int i = k; i < nums.length; i++) {
             sum += nums[i] - nums[i - k];
             maxSum = Math.max(maxSum, sum);
         }
-        return maxSum / k;
+        return (double) maxSum / k;
     }
 }
